rename the main class, and add grpc config property
