+++ fotile-grpc-invoker/src/main/java/com/fotile/grpc/invoker/InvokerApplication.java
@@ -1,16 +1,16 @@
-package com.fotile.grpc.server;
-
+package com.fotile.grpc.invoker;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class Application {
+public class InvokerApplication {
 
 
     public static void main(String[] args) {
 
-        SpringApplication.run(Application.class, args);
+        SpringApplication.run(InvokerApplication.class, args);
+
     }
 
 }
