+++ fotile-grpc-server/src/main/java/com/fotile/grpc/server/GrpcServerApplication.java
@@ -1,16 +1,16 @@
-package com.fotile.grpc.invoker;
+package com.fotile.grpc.server;
+
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class Application {
+public class GrpcServerApplication {
 
 
     public static void main(String[] args) {
 
-        SpringApplication.run(Application.class, args);
-
+        SpringApplication.run(GrpcServerApplication.class, args);
     }
 
 }
