+++ fotile-grpc-server/src/main/resources/application.yml
@@ -1,0 +1,10 @@
+server:
+  port: 9091
+spring:
+  application:
+    name: fotile-grpc-server
+
+grpc:
+  server:
+    address: 127.0.0.1
+    port: 9092
